+++ common/src/main/java/com/github/yajatkaul/mega_showdown/mixin/client/PokemonRendererMixin.java
@@ -17,6 +17,8 @@ import com.github.yajatkaul.mega_showdown.render.LayerDataLoader;
 import com.github.yajatkaul.mega_showdown.render.layerEntities.DynamaxCloudsLayer;
 import com.github.yajatkaul.mega_showdown.render.layerEntities.TeraHatsLayer;
 import com.github.yajatkaul.mega_showdown.render.layerEntities.states.TeraCrystalState;
+import com.github.yajatkaul.mega_showdown.utils.duck.cobblemon.interfaces.ClientBattleDuck;
+import com.github.yajatkaul.mega_showdown.utils.duck.cobblemon.interfaces.PokemonEntityDuck;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.math.Axis;
@@ -52,18 +54,7 @@ public class PokemonRendererMixin {
     @Unique
     private final ResourceLocation mega_showdown$teraCrystalPoserId = ResourceLocation.fromNamespaceAndPath("cobblemon", "terastal_transformation");
     @Unique
-    public final TeraCrystalState mega_showdown$teraCrystalState = new TeraCrystalState();
-    @Unique
     private final Set<String> mega_showdown$teraCrystalAspects = new HashSet<>();
-    @Unique
-    private boolean mega_showdown$teraCrystalPlayed = false;
-    @Unique
-    private boolean mega_showdown$teraCrystalPass = false;
-    //Timershit
-    @Unique
-    private double mega_showdown$animCrystalSeconds = 0.0;
-    @Unique
-    private long mega_showdown$lastCrystalTimeNs = -1L;
     //
 
     @Inject(method = "<init>", at = @At(value = "RETURN"))
@@ -81,7 +72,10 @@ public class PokemonRendererMixin {
         Optional<String> aspect = pokemon.getAspects().stream()
                 .filter(a -> a.startsWith("msd:tera_")).findFirst();
 
-        if (tera_play && (!mega_showdown$teraCrystalPlayed || mega_showdown$teraCrystalPass) && aspect.isPresent()) {
+
+        boolean mega_showdown$teraCrystalPlayed = ((PokemonEntityDuck) (Object) entity).mega_showdown$isTeraCrystalPlayed();
+        boolean mega_showdown$teraCrystalPass = ((PokemonEntityDuck) (Object) entity).mega_showdown$isTeraCrystalPass();
+        if (tera_play && (!mega_showdown$teraCrystalPlayed || mega_showdown$teraCrystalPass)) {
             mega_showdown$renderTeraCrystals(
                     entity,
                     pokemon,
@@ -110,36 +104,42 @@ public class PokemonRendererMixin {
 
     @Unique
     private void mega_showdown$renderTeraCrystals(PokemonEntity entity, Pokemon pokemon, PokemonClientDelegate clientDelegate, PoseStack poseStack, MultiBufferSource buffer, int packedLight) {
+        PokemonEntityDuck duck = ((PokemonEntityDuck) (Object) entity);
+        long mega_showdown$lastCrystalTimeNs = duck.mega_showdown$getLastCrystalTimeNs();
+        double mega_showdown$animCrystalSeconds = duck.mega_showdown$getAnimCrystalSeconds();
+
         if (!Minecraft.getInstance().isPaused()) {
             long now = System.nanoTime();
 
             if (mega_showdown$lastCrystalTimeNs != -1L) {
                 double deltaSeconds = (now - mega_showdown$lastCrystalTimeNs) / 1_000_000_000.0;
-                mega_showdown$animCrystalSeconds += deltaSeconds;
+                duck.mega_showdown$setAnimCrystalSeconds(mega_showdown$animCrystalSeconds + deltaSeconds);
             }
 
-            mega_showdown$lastCrystalTimeNs = now;
+            duck.mega_showdown$setLastCrystalTimeNs(now);
         } else {
-            mega_showdown$lastCrystalTimeNs = System.nanoTime();
+            duck.mega_showdown$setLastCrystalTimeNs(System.nanoTime());
         }
 
         float mega_showdown$teraCrystalDuration = new BedrockActiveAnimation(
                 BedrockAnimationRepository.INSTANCE.getAnimation("terastal_transformation", "animation.terastal_transformation.transform")
         ).getDuration();
 
+        TeraCrystalState mega_showdown$teraCrystalState = ((PokemonEntityDuck) (Object) entity).mega_showdown$getTeraCrystalState();
+
         if (mega_showdown$teraCrystalState.getAnimationSeconds() >= mega_showdown$teraCrystalDuration) {
-            mega_showdown$teraCrystalPlayed = true;
-            mega_showdown$teraCrystalPass = false;
-            mega_showdown$animCrystalSeconds = 0.0;
-            mega_showdown$lastCrystalTimeNs = -1L;
+            duck.mega_showdown$setTeraCrystalPlayed(true);
+            duck.mega_showdown$setTeraCrystalPass(false);
+            duck.mega_showdown$setAnimCrystalSeconds(0.0);
+            duck.mega_showdown$setLastCrystalTimeNs(-1L);
             mega_showdown$teraCrystalState.resetAnimation();
             entity.after(3f, () -> {
-                mega_showdown$teraCrystalPlayed = false;
+                duck.mega_showdown$setTeraCrystalPlayed(false);
                 return Unit.INSTANCE;
             });
             return;
         } else if (mega_showdown$teraCrystalState.getAnimationSeconds() >= mega_showdown$teraCrystalDuration - 0.3) {
-            mega_showdown$teraCrystalPass = true;
+            duck.mega_showdown$setTeraCrystalPass(true);
         }
 
         float ticks = (float) (mega_showdown$animCrystalSeconds * 20f);