+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.5.1+1.7.1-b-7
 - fixed dynamax scale stuck issue
+- fixed tera crystal state (independent, you dont have to know what that means dw)